Some refactoring and tweaks
